+++ src/Resources/app/storefront/src/main.js
@@ -7,6 +7,7 @@ import PayonePaymentRatepayInstallment from "./ratepay-installment/payone-paymen
 import PayonePaymentKlarna from "./klarna/payone-payment.klarna";
 import PayonePaymentPayPalV2 from "./paypal-v2/payone-payment.paypal-v2";
 import "./amazon-pay";
+import PayoneGooglePayPlugin from "./google-pay/payone-payment.google-pay";
 
 const PluginManager = window.PluginManager;
 
@@ -18,6 +19,7 @@ PluginManager.register('PayonePaymentApplePay', PayonePaymentApplePay, '[data-pa
 PluginManager.register('PayonePaymentRatepayInstallment', PayonePaymentRatepayInstallment, '[data-is-payone-ratepay-installment]');
 PluginManager.register('PayonePaymentKlarna', PayonePaymentKlarna, '[data-payone-payment-klarna]');
 PluginManager.register('PayonePaymentPayPalV2', PayonePaymentPayPalV2, '[data-payone-payment-pay-pal-v2-options]');
+PluginManager.register('PayoneGooglePayPlugin', PayoneGooglePayPlugin, '[data-payone-google-pay]');
 
 if (module.hot) {
     module.hot.accept();